Bug Fix: Shift left by 24 with wrong data type
Thanks @darudev for letting me know =).

+++ src/sha1.c
@@ -247,10 +247,10 @@ void SHA1ProcessMessageBlock(SHA1Context *context)
      */
     for(t = 0; t < 16; t++)
     {
-        W[t] = context->Message_Block[t * 4] << 24;
-        W[t] |= context->Message_Block[t * 4 + 1] << 16;
-        W[t] |= context->Message_Block[t * 4 + 2] << 8;
-        W[t] |= context->Message_Block[t * 4 + 3];
+        W[t]  = (uint32_t)context->Message_Block[t * 4] << 24;
+        W[t] |= (uint32_t)context->Message_Block[t * 4 + 1] << 16;
+        W[t] |= (uint32_t)context->Message_Block[t * 4 + 2] << 8;
+        W[t] |= (uint32_t)context->Message_Block[t * 4 + 3];
     }
 
     for(t = 16; t < 80; t++)
